cleanup sites code
